requirements.txt
about-time==4.2.1
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp_socks==0.10.1
aiosignal==1.3.2
alive-progress==3.2.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
black==25.1.0
blinker==1.9.0
bs4==0.0.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
executing==2.2.0
flake8==7.2.0
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.5.0
future==1.0.0
future-annotations==1.0.0
grapheme==0.6.0
h11==0.14.0
h8mail==2.5.6
holehe==1.61
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.0.5
kiwisolver==1.4.8
lolcat==1.4
lxml==5.3.2
maigret @ file:///home/ncacord/Auton-OSINT/tools/maigret
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mock==5.2.0
multidict==6.4.2
mypy-extensions==1.0.0
networkx==2.8.8
numpy==2.2.4
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.2
param==2.2.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==8.10.2
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.6.0
pycodestyle==2.13.0
pycountry==24.6.1
pycparser==2.22
pyflakes==3.3.2
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.3.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-socks==2.7.1
pyvis==0.3.2
pyviz_comms==3.0.4
PyYAML==6.0.2
qrcode==8.1
reportlab==4.3.1
requests==2.32.3
requests-futures==1.0.2
requests-toolbelt==1.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
socid-extractor==0.0.27
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
stem==1.8.2
svglib==1.5.1
termcolor==3.0.1
tinycss2==1.4.0
tokenize_rt==6.1.0
toml==0.10.2
torrequest==0.1.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
typing_extensions==4.13.1
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
whl==0.0.4
whois==1.20240129.2
xhtml2pdf==0.2.17
XMind==1.2.0
yarl==1.19.0
